+++ src/Symfony/Component/Routing/Tests/Loader/XmlFileLoaderTest.php
@@ -53,6 +53,7 @@ class XmlFileLoaderTest extends TestCase
         $this->assertEquals(['https'], $route->getSchemes());
         $this->assertEquals('context.getMethod() == "GET"', $route->getCondition());
         $this->assertTrue($route->getDefault('_stateless'));
+        $this->assertTrue($route->getDefault('_static_generation'));
     }
 
     public function testLoadWithNamespacePrefix()
@@ -105,6 +106,7 @@ class XmlFileLoaderTest extends TestCase
         $this->assertSame('en', $defaultsRoute->getDefault('_locale'));
         $this->assertSame('html', $defaultsRoute->getDefault('_format'));
         $this->assertTrue($defaultsRoute->getDefault('_stateless'));
+        $this->assertTrue($defaultsRoute->getDefault('_static_generation'));
     }
 
     public function testLoadingImportedRoutesWithDefaults()
@@ -119,10 +121,12 @@ class XmlFileLoaderTest extends TestCase
         $localeRoute->setDefault('_locale', 'g_locale');
         $localeRoute->setDefault('_format', 'g_format');
         $localeRoute->setDefault('_stateless', true);
+        $localeRoute->setDefault('_static_generation', true);
         $expectedRoutes->add('two', $formatRoute = new Route('/defaults/two'));
         $formatRoute->setDefault('_locale', 'g_locale');
         $formatRoute->setDefault('_format', 'g_format');
         $formatRoute->setDefault('_stateless', true);
+        $formatRoute->setDefault('_static_generation', true);
         $formatRoute->setDefault('specific', 'imported');
 
         $expectedRoutes->addResource(new FileResource(__DIR__.'/../Fixtures/imported-with-defaults.xml'));