+++ src/Symfony/Component/Routing/Tests/Loader/YamlFileLoaderTest.php
@@ -99,6 +99,7 @@ class YamlFileLoaderTest extends TestCase
         $this->assertEquals(['https'], $route->getSchemes());
         $this->assertEquals('context.getMethod() == "GET"', $route->getCondition());
         $this->assertTrue($route->getDefault('_stateless'));
+        $this->assertTrue($route->getDefault('_static_generation'));
     }
 
     public function testLoadWithResource()
@@ -244,6 +245,7 @@ class YamlFileLoaderTest extends TestCase
         $this->assertSame('en', $defaultsRoute->getDefault('_locale'));
         $this->assertSame('html', $defaultsRoute->getDefault('_format'));
         $this->assertTrue($defaultsRoute->getDefault('_stateless'));
+        $this->assertTrue($defaultsRoute->getDefault('_static_generation'));
     }
 
     public function testLoadingImportedRoutesWithDefaults()
@@ -258,10 +260,12 @@ class YamlFileLoaderTest extends TestCase
         $localeRoute->setDefault('_locale', 'g_locale');
         $localeRoute->setDefault('_format', 'g_format');
         $localeRoute->setDefault('_stateless', true);
+        $localeRoute->setDefault('_static_generation', true);
         $expectedRoutes->add('two', $formatRoute = new Route('/defaults/two'));
         $formatRoute->setDefault('_locale', 'g_locale');
         $formatRoute->setDefault('_format', 'g_format');
         $formatRoute->setDefault('_stateless', true);
+        $formatRoute->setDefault('_static_generation', true);
         $formatRoute->setDefault('specific', 'imported');
 
         $expectedRoutes->addResource(new FileResource(__DIR__.'/../Fixtures/imported-with-defaults.yml'));